ASoC: rockchip: i2s-tdm: Compatible for deprecated property 'clk-trcm'
In kernel 6.1 and later versions, the property 'clk-trcm' has been updated:

  rockchip,clk-trcm -> rockchip,trcm-sync-tx/rx-only

Not all products' DT have been updated to use the new property. Therefore,
still support for old property for now, and a prompt is logged to suggest
to adopt the new property.

The old property will be removed in a future version.

Log:

  Deprecated property 'rockchip,clk-trcm', please use 'rockchip,trcm-sync-tx/rx-only' in DT

Signed-off-by: Sugar Zhang <sugar.zhang@rock-chips.com>
Change-Id: I594ba932486440d6eefc65abb54516d530ad090b